feat(settings): nest settings as a project tab; drop pickers + dead routes
Settings now lives at /project/[projectId]/settings (inside the project layout,
so the project sidebar stays visible) as a side-nav item again — not a standalone
page. The secondary nav (workspace + project sections) renders in the project
layout's content slot, styled like the old settings panel (bg-background, own
SidebarProvider, header with sidebar toggle + breadcrumb).

- Remove the workspace/project picker dropdowns: settings implicitly targets the
  current project/workspace (shown as static labels), which also removes the need
  for any target-decoupling machinery.
- Move Settings from the sidebar footer back into the project side nav.
- Clean up detour cruft from the unshipped /settings rework:
  - delete the dead, unimported old panel components/settings/settings.tsx
  - delete the orphaned bare /settings and 2-segment /settings/[ws]/[proj] routes
  - keep /settings/[ws] as the workspace->first-project settings resolver
    (reached by the /workspace shim + workspace-create fallback)
